requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.37.0
chromadb>=0.5.4
pydantic>=2.8.0
PyYAML>=6.0.1
python-docx>=1.1.2
pypdf>=4.3.1
pandas>=2.2.2
requests>=2.32.3
python-multipart>=0.0.9
edge-tts>=6.1.18
